cleanup a la knut
