+++ tests/Digdir.Domain.Dialogporten.WebAPI.E2E.Tests/Features/V1/ServiceOwner/Dialogs/Queries/Search/SearchDialogTests.cs
@@ -1,5 +1,4 @@
 using System.Net;
-using Altinn.ApiClients.Dialogporten.Features.V1;
 using AwesomeAssertions;
 using Digdir.Domain.Dialogporten.Domain.Parties;
 using Digdir.Library.Dialogporten.E2E.Common;
@@ -10,45 +9,6 @@ namespace Digdir.Domain.Dialogporten.WebAPI.E2E.Tests.Features.V1.ServiceOwner.D
 [Collection(nameof(WebApiTestCollectionFixture))]
 public class SearchDialogTests(WebApiE2EFixture fixture) : E2ETestBase<WebApiE2EFixture>(fixture)
 {
-    [E2ETheory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task Should_Filter_By_ServiceOwnerLabels(bool withEndUserId)
-    {
-        // Arrange
-        var sentinelLabel = Guid.NewGuid().ToString();
-        var controlLabel = Guid.NewGuid().ToString();
-
-        var dialogId = await Fixture.ServiceownerApi.CreateSearchDialogAsync(sentinelLabel);
-        var controlDialogId = await Fixture.ServiceownerApi.CreateSearchDialogAsync(controlLabel);
-
-        // Act
-        var searchResult = await E2ERetryPolicies.RetryUntilAsync(
-            ct => Fixture.ServiceownerApi.V1ServiceOwnerDialogsQueriesSearchDialog(new()
-            {
-                ServiceOwnerLabels = [sentinelLabel],
-                Party = withEndUserId
-                    ? [E2EConstants.DefaultParty]
-                    : null!,
-                EndUserId = withEndUserId
-                    ? E2EConstants.DefaultParty
-                    : null!
-            }, ct),
-            isSuccessful: response =>
-                response.Content is { Items.Count: 1 } content
-                && content.Items.Single().Id == dialogId
-                && content.Items.All(x => x.Id != controlDialogId),
-            degradationMessage: "Search indexing speed is degraded.");
-
-        // Assert
-        searchResult.ShouldHaveStatusCode(HttpStatusCode.OK);
-        searchResult.Content.Should().NotBeNull();
-
-        var content = searchResult.Content;
-        content.Items.Should().ContainSingle();
-        content.Items.Single().Id.Should().Be(dialogId);
-    }
-
     [E2EFact]
     public async Task Should_Return_Items_On_Simple_List()
     {
@@ -355,46 +315,6 @@ public class SearchDialogTests(WebApiE2EFixture fixture) : E2ETestBase<WebApiE2E
         var content = searchResult.Content;
         content.Items.Should().ContainSingle(x => x.Id == dialogId);
         content.Items.Should().NotContain(x => x.Id == controlDialogId);
-    }
-
-    [E2ETheory]
-    [InlineData(true)]
-    [InlineData(false)]
-    public async Task Should_Filter_By_ServiceResource(bool withEndUserId)
-    {
-        // Arrange
-        var sentinelLabel = Guid.NewGuid().ToString();
-
-        var dialogId = await Fixture.ServiceownerApi.CreateSearchDialogAsync(sentinelLabel, dialog =>
-        {
-            dialog.ServiceResource = E2EConstants.AlternateServiceResource;
-        });
-
-        var controlDialogId = await Fixture.ServiceownerApi.CreateSearchDialogAsync(sentinelLabel);
-
-        // Act
-        var searchResult = await E2ERetryPolicies.RetryUntilAsync(
-            ct => Fixture.ServiceownerApi.V1ServiceOwnerDialogsQueriesSearchDialog(new()
-            {
-                ServiceOwnerLabels = [sentinelLabel],
-                ServiceResource = [E2EConstants.AlternateServiceResource],
-                EndUserId = withEndUserId ? E2EConstants.DefaultParty : null!
-            }, ct),
-            isSuccessful: response =>
-                response.Content is { Items.Count: 1 } content
-                && content.Items.Single().Id == dialogId
-                && content.Items.All(x => x.Id != controlDialogId),
-            degradationMessage: "Search indexing speed is degraded.");
-
-        // Assert
-        searchResult.ShouldHaveStatusCode(HttpStatusCode.OK);
-        searchResult.Content.Should().NotBeNull();
-
-        var content = searchResult.Content;
-        content.Items.Should().ContainSingle();
-        content.Items.Single().Id.Should().Be(dialogId);
-        content.Items.Single().ServiceResource.Should()
-            .Be(E2EConstants.AlternateServiceResource);
     }
 
     [E2EFact]
